add changes according to code review
