+++ Project X/Welcome Module/WelcomeInteractor/WelcomeInteractorIO.swift
@@ -1,0 +1,15 @@
+//
+//  WelcomeInteractorIO.swift
+//  Project X
+//
+//  Created by Nataliya Lazouskaya on 14.09.22.
+//
+
+import Foundation
+
+protocol WelcomeInteractorInputProtocol {
+   init(presenter: WelcomeInteractorOutputProtocol)
+}
+
+protocol WelcomeInteractorOutputProtocol: AnyObject {
+}
